Add a per-language option to hide the banner
By default, we set the "hide-banner" class on the body element for any
language other than "en". We mark "hide-banner" for translation, so once
the banner content has been translated in a particular language, they
can add a translation for "hide-banner" that replaces it with an empty
string.

This change requires additional support in the CSS provided by the
creativecommons.org repository.