This is a big scary commit that removes cc.engine.
cc.engine's history from this repo has been moved back to the main cc.engine
repo and the scripts that fetch the Python code have been changed to reflect
this.

Nothing has been lost and the code will continue to work as before once
the new directory structure is recreated on the server.